+++ src/languages/csp.js
@@ -11,8 +11,8 @@ export default function(hljs) {
   return {
     name: 'CSP',
     case_insensitive: false,
-    lexemes: '[a-zA-Z][a-zA-Z0-9_-]*',
     keywords: {
+      $pattern: '[a-zA-Z][a-zA-Z0-9_-]*',
       keyword: 'base-uri child-src connect-src default-src font-src form-action ' +
         'frame-ancestors frame-src img-src media-src object-src plugin-types ' +
         'report-uri sandbox script-src style-src',