+++ src/languages/gams.js
@@ -10,14 +10,14 @@
 
 export default function (hljs) {
   var KEYWORDS = {
-    'keyword':
+    keyword:
       'abort acronym acronyms alias all and assign binary card diag display ' +
       'else eq file files for free ge gt if integer le loop lt maximizing ' +
       'minimizing model models ne negative no not option options or ord ' +
       'positive prod put putpage puttl repeat sameas semicont semiint smax ' +
       'smin solve sos1 sos2 sum system table then until using while xor yes',
-    'literal': 'eps inf na',
-    'built-in':
+    literal: 'eps inf na',
+    built_in:
       'abs arccos arcsin arctan arctan2 Beta betaReg binomial ceil centropy ' +
       'cos cosh cvPower div div0 eDist entropy errorf execSeed exp fact ' +
       'floor frac gamma gammaReg log logBeta logGamma log10 log2 mapVal max ' +