+++ src/lib/modes.js
@@ -63,7 +63,7 @@ export const COMMENT = function(begin, end, inherits) {
   mode.contains.push(PHRASAL_WORDS_MODE);
   mode.contains.push({
     className: 'doctag',
-    begin: '(?:TODO|FIXME|NOTE|BUG|XXX):',
+    begin: '(?:TODO|FIXME|NOTE|BUG|OPTIMIZE|HACK|XXX):',
     relevance: 0
   });
   return mode;