+++ tools/build_config.js
@@ -29,7 +29,8 @@ module.exports = {
         },
         output: {
           name: "hljs",
-          format: "umd",
+          format: "iife",
+          footer: "if (typeof exports === 'object' && typeof module !== 'undefined') { module.exports = hljs; }",
           interop: false,
         }
       },